Use unsigned data types for numeric limits
Fixes issues with conversion of EO Int types > int.MaxValue

+++ EOLib/Data/EoWriter.cs
@@ -250,9 +250,9 @@ public sealed class EoWriter
         return result;
     }
 
-    private static void CheckNumberSize(int number, int max)
+    private static void CheckNumberSize(int number, uint max)
     {
-        if ((uint)number > (uint)max)
+        if ((uint)number > max)
         {
             throw new ArgumentException($"Value {number} exceeds maximum of {max}.");
         }
